241111 13:39 황진- 나눔게시판 목록, 글 삭제 작업

+++ src/main/java/com/example/moamz/service/community/sharing/SharingBoardService.java
@@ -118,8 +118,18 @@ public class SharingBoardService {
 
 
     // 3️⃣ 게시글 삭제 메서드
-    public void removeSharing(Long postId) {
-        sharingBoardMapper.deleteSharing(postId);
+    public Boolean removeSharing(Long userCode, Long postId) {
+
+        // 게시글 작성자
+        Long postWriter = sharingBoardMapper.selectWriter(postId);
+
+        // 삭제를 요청한 userCode와 게시글 작성자 userCode가 일치할 때만 삭제 가능
+        if(postWriter.equals(userCode)) {
+            sharingBoardMapper.deleteSharing(postId);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     // 4️⃣ 게시글 목록 가져오기 메서드
